test: reorganize test struct for attribute traits
